refactor(notion): split Annotaion from RichText
